requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
ytmusicapi>=1.7.0
httpx>=0.27.0
python-dotenv>=1.0.0
jinja2>=3.1.3
playwright>=1.44.0
browser-cookie3>=0.19.1
